+++ src/main/java/roomescape/dao/JdbcReservationTimeDao.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Component;
 import roomescape.domain.ReservationTime;
-import roomescape.mapper.ReservationTimeMapper;
+import roomescape.domain.mapper.ReservationTimeMapper;
 
 @Component
 public class JdbcReservationTimeDao implements ReservationTimeDao {
