+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -8,13 +8,14 @@ import java.time.LocalTime;
 import java.util.List;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import roomescape.application.dto.ReservationRequest;
+import roomescape.application.dto.ReservationResponse;
+import roomescape.application.dto.ReservationTimeResponse;
+import roomescape.application.service.ReservationService;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.dao.ThemeDao;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
-import roomescape.dto.ReservationTimeResponse;
 import roomescape.service.fake_dao.FakeReservationDao;
 import roomescape.service.fake_dao.FakeReservationTimeDao;
 import roomescape.service.fake_dao.FakeThemeDao;
